+++ lib/Predis/Pipeline/StandardExecutor.php
@@ -13,7 +13,9 @@ namespace Predis\Pipeline;
 
 use SplQueue;
 use Predis\ResponseErrorInterface;
+use Predis\ResponseObjectInterface;
 use Predis\ServerException;
+use Predis\Command\CommandInterface;
 use Predis\Connection\ConnectionInterface;
 use Predis\Connection\ReplicationConnectionInterface;
 
@@ -51,6 +53,27 @@ class StandardExecutor implements PipelineExecutorInterface
     }
 
     /**
+     * Handles a response object.
+     *
+     * @param ConnectionInterface $connection
+     * @param CommandInterface $command
+     * @param ResponseObjectInterface $response
+     * @return mixed
+     */
+    protected function onResponseObject(ConnectionInterface $connection, CommandInterface $command, ResponseObjectInterface $response)
+    {
+        if ($response instanceof ResponseErrorInterface) {
+            return $this->onResponseError($connection, $response);
+        }
+
+        if ($response instanceof \Iterator) {
+            return $command->parseResponse(iterator_to_array($response));
+        }
+
+        return $response;
+    }
+
+    /**
      * Handles -ERR responses returned by Redis.
      *
      * @param ConnectionInterface $connection The connection that returned the error.
@@ -58,6 +81,10 @@ class StandardExecutor implements PipelineExecutorInterface
      */
     protected function onResponseError(ConnectionInterface $connection, ResponseErrorInterface $response)
     {
+        if (!$this->exceptions) {
+            return $response;
+        }
+
         // Force disconnection to prevent protocol desynchronization.
         $connection->disconnect();
         $message = $response->getMessage();
@@ -72,7 +99,6 @@ class StandardExecutor implements PipelineExecutorInterface
     {
         $size = count($commands);
         $values = array();
-        $exceptions = $this->exceptions;
 
         $this->checkConnection($connection);
 
@@ -81,13 +107,14 @@ class StandardExecutor implements PipelineExecutorInterface
         }
 
         for ($i = 0; $i < $size; $i++) {
-            $response = $connection->readResponse($commands->dequeue());
+            $command = $commands->dequeue();
+            $response = $connection->readResponse($command);
 
-            if ($response instanceof ResponseErrorInterface && $exceptions === true) {
-                $this->onResponseError($connection, $response);
+            if ($response instanceof ResponseObjectInterface) {
+                $values[$i] = $this->onResponseObject($connection, $command, $response);
+            } else {
+                $values[$i] = $command->parseResponse($response);
             }
-
-            $values[$i] = $response instanceof \Iterator ? iterator_to_array($response) : $response;
         }
 
         return $values;